+++ packages/ethereum-contracts/contracts/interfaces/superfluid/Definitions.sol
@@ -160,6 +160,24 @@ library BatchOperation {
      */
     uint32 constant internal OPERATION_TYPE_ERC777_SEND = 3;
     /**
+     * @dev ERC20.increaseAllowance batch operation type
+     *
+     * Call spec:
+     * ISuperToken(target).operationIncreaseAllowance(
+     *     abi.decode(data, (address account, address spender, uint256 addedValue))
+     * )
+     */
+    uint32 constant internal OPERATION_TYPE_ERC20_INCREASE_ALLOWANCE = 4;
+    /**
+     * @dev ERC20.decreaseAllowance batch operation type
+     *
+     * Call spec:
+     * ISuperToken(target).operationDecreaseAllowance(
+     *     abi.decode(data, (address account, address spender, uint256 subtractedValue))
+     * )
+     */
+    uint32 constant internal OPERATION_TYPE_ERC20_DECREASE_ALLOWANCE = 5;
+    /**
      * @dev SuperToken.upgrade batch operation type
      *
      * Call spec: