+++ packages/ethereum-contracts/CHANGELOG.md
@@ -18,6 +18,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 - bump solc to 0.8.18
+- New ACL functions: `increaseFlowRateAllowance` and `decreaseFlowRateAllowance` in `ConstantFlowAgreementV1.sol`
+- Support for `superToken.increaseAllowance` and `superToken.decreaseAllowance` in `batchCall` in `Superfluid.sol`
+
 ### Breaking
 - `SuperTokenFactory` contract no longer takes `SuperTokenHelper` contract in its constructor
   - Migration: Pass in a deployed `SuperToken` (logic) contract address to `SuperTokenFactory` constructor instead